better organization of layout.py
