+++ descwl_shear_sims/layout/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .layout import Layout
